Git move (#17)
* Added git-move command

* Update git-commit to use git-move to copy files

* Added git-move as an option to :better-export

* Moved $developerOptions and $gitOptions variable declaration

* git-init wouldn’t work as intended

* git-commit no longer moves files

The git-commit and git-move actions should be 100% separated, as they represent separate operations. With the new -m param to better-export, if this behaviour is needed it should be explicitly called.

* Switched position for git-move and build-release

It’s better to build the release before moving to the git repo, in case any errors occur in the build process. Errors should be dealt with before placing the files under version control.